requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.0
google-cloud-vision==3.4.4
pytest==7.4.3
httpx==0.25.1 